Relocate parser tests
