+++ test/utils/parsers/android.test.ts
@@ -1,6 +1,6 @@
-import { executeFileDiff } from '../../../../src/utils/diff/fileDiff'
+import { executeFileDiff } from '../../../src/utils/diff/fileDiff'
 import * as path from 'node:path'
-import { parseFiles } from '../../../../src/utils/diff/parse'
+import { parseFiles } from '../../../src/utils/diff/parse'
 import { expect } from '@oclif/test'
 
 describe('android', () => {
@@ -43,7 +43,7 @@ describe('android', () => {
         }
     ]
     it('identifies the correct variable usages in the Android sample diff', () => {
-        const parsedDiff = executeFileDiff(path.join(__dirname, '../samples/android'))
+        const parsedDiff = executeFileDiff(path.join(__dirname, '../diff/samples/android'))
         const results = parseFiles(parsedDiff)
 
         expect(results).to.deep.equal({
