+++ test/utils/parsers/csharp.test.ts
@@ -1,6 +1,6 @@
-import { executeFileDiff } from '../../../../src/utils/diff/fileDiff'
+import { executeFileDiff } from '../../../src/utils/diff/fileDiff'
 import * as path from 'node:path'
-import { parseFiles } from '../../../../src/utils/diff/parse'
+import { parseFiles } from '../../../src/utils/diff/parse'
 import { expect } from '@oclif/test'
 
 describe('csharp', () => {
@@ -43,7 +43,7 @@ describe('csharp', () => {
         }
     ]
     it('identifies the correct variable usages in the C# sample diff', () => {
-        const parsedDiff = executeFileDiff(path.join(__dirname, '../samples/csharp'))
+        const parsedDiff = executeFileDiff(path.join(__dirname, '../diff/samples/csharp'))
         const results = parseFiles(parsedDiff)
 
         expect(results).to.deep.equal({
