+++ test/utils/parsers/java.test.ts
@@ -1,6 +1,6 @@
-import { executeFileDiff } from '../../../../src/utils/diff/fileDiff'
+import { executeFileDiff } from '../../../src/utils/diff/fileDiff'
 import * as path from 'node:path'
-import { parseFiles } from '../../../../src/utils/diff/parse'
+import { parseFiles } from '../../../src/utils/diff/parse'
 import { expect } from '@oclif/test'
 
 describe('java', () => {
@@ -31,7 +31,7 @@ describe('java', () => {
         }
     ]
     it('identifies the correct variable usages in the Java sample diff', () => {
-        const parsedDiff = executeFileDiff(path.join(__dirname, '../samples/java'))
+        const parsedDiff = executeFileDiff(path.join(__dirname, '../diff/samples/java'))
         const results = parseFiles(parsedDiff)
 
         expect(results).to.deep.equal({
