+++ test/utils/parsers/php.test.ts
@@ -1,6 +1,6 @@
-import { executeFileDiff } from '../../../../src/utils/diff/fileDiff'
+import { executeFileDiff } from '../../../src/utils/diff/fileDiff'
 import * as path from 'node:path'
-import { parseFiles } from '../../../../src/utils/diff/parse'
+import { parseFiles } from '../../../src/utils/diff/parse'
 import { expect } from '@oclif/test'
 
 describe('php', () => {
@@ -37,7 +37,7 @@ describe('php', () => {
         }
     ]
     it('identifies the correct variable usages in the PHP sample diff', () => {
-        const parsedDiff = executeFileDiff(path.join(__dirname, '../samples/php'))
+        const parsedDiff = executeFileDiff(path.join(__dirname, '../diff/samples/php'))
         const results = parseFiles(parsedDiff)
 
         expect(results).to.deep.equal({
