+++ test/utils/parsers/react.test.ts
@@ -1,6 +1,6 @@
-import { executeFileDiff } from '../../../../src/utils/diff/fileDiff'
+import { executeFileDiff } from '../../../src/utils/diff/fileDiff'
 import * as path from 'node:path'
-import { parseFiles } from '../../../../src/utils/diff/parse'
+import { parseFiles } from '../../../src/utils/diff/parse'
 import { expect } from '@oclif/test'
 
 describe('react', () => {
@@ -26,7 +26,7 @@ describe('react', () => {
         }
     ]
     it('identifies the correct variable usages in the React sample diff', () => {
-        const parsedDiff = executeFileDiff(path.join(__dirname, '../samples/react'))
+        const parsedDiff = executeFileDiff(path.join(__dirname, '../diff/samples/react'))
         const results = parseFiles(parsedDiff)
 
         expect(results).to.deep.equal({
